[modularize] Separate exceptions from `Worker`.
It is because many modules will share them later.
Separate it can avoid circular reference.
